Split the self.ds is None and ds != 'r+' in two
Reasoning: if the bug is that self.ds is set to None for some reason,
an error message about read-only mode will be misleading.

I imagine this might happen if an incorrect file path is passed.

Closed files are treated as read-only

Error messages now include simple description of where they occurred.